Made some small (non-breaking) changes that now mean that the tests can all be run using pytest
Running "pytest ." in the "unittests" directory now works. There is a lot
of time to collect the tests as we don't properly separate tests from
reading in data. I will clean these up over time. Also some small
issues with running the locale tests. But, overall, a pytest
interface and run is much nicer and cleaner. It will also let us
mark tests as "slow", so that it is possible to run a quicker
set of tests while developing.

(note that the symbolic link of the io directory is needed as pytest
runs everything in the unittests directory, while the tests
were written expecting to run in their own directories. Adding
the symbolic link means that "../io" works in both cases)